requirements.txt
Appium-Python-Client
percy-appium
requests==2.*